backend/requirements.txt
annotated-types==0.7.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
blinker==1.9.0
blis==1.3.0
camel-converter==4.0.1
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
compressed-rtf==1.0.7
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
docx2txt==0.9
ebcdic==1.1.1
et_xmlfile==2.0.0
extract-msg==0.28.7
filelock==3.18.0
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
ijson==3.3.0
IMAPClient==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meilisearch==0.34.1
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy==2.2.5
olefile==0.47
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20250416
pillow==11.2.1
preshed==3.0.9
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyPDF2==3.0.1
PySimpleGUI==5.0.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-pptx==0.6.23
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.12.0
smart-open==7.1.0
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
srsly==2.5.1
sympy==1.14.0
textract==1.6.5
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==1.2.0
XlsxWriter==3.2.3
